requirements.txt
langchain==0.3.13
langchain-ollama==0.2.6
langchain-community==0.3.13
langchain-chroma==0.2.0
chromadb==0.5.23
streamlit==1.41.0
python-dotenv==1.0.1
pypdf==5.1.0
tiktoken==0.8.0